fix(finance): store currency amounts as numeric(14,2), not double precision
All 26 monetary columns were double precision. Binary floating point
cannot represent most decimal money values, which is not theoretical
here:

  SELECT sum(v) FROM (ten rows of 0.1)
    double precision -> 0.9999999999999999 , (sum = 1.0) IS FALSE
    numeric(14,2)    -> 1.00               , (sum = 1.0) IS TRUE

That false is the shape of the comparison in
bank_statement_integrity._payment_matches and in fns/receipt_verify,
where a payment is reconciled against a bank or fiscal claim by amount.
Budget totals were likewise order-dependent, and a written amount could
persist as 1234.5600000000001.

w24 converts the columns with round(col::numeric, 2): each stored double
is rounded once, deterministically, to the value it always meant.

asdecimal=False is deliberate and load-bearing: the ORM keeps returning
float, so no service arithmetic, response schema or mobile type changes.
This fixes storage, SQL aggregation and SQL comparison. Moving the
finance services to Decimal arithmetic is a separate follow-up and is
not claimed here.

Measurements (m², %, quantities, coefficients, ratings, pin coordinates)
stay double precision on purpose — converting them would change quantity
semantics without cause. A test asserts that boundary in both directions.

Depends on w23 (fix/scope-indexes) for the migration chain.
No deletions.

+++ backend/app/models/project_profile.py
@@ -7,6 +7,8 @@ column or a second source of truth.
 from __future__ import annotations
 
 from sqlalchemy import Float
+
+from app.models.money import MONEY
 from sqlalchemy.orm import mapped_column
 
 from app.models.entities import Project
@@ -21,7 +23,7 @@ def register_project_profile_fields() -> None:
         setattr(
             Project,
             "customer_budget",
-            mapped_column("customer_budget", Float, nullable=True),
+            mapped_column("customer_budget", MONEY, nullable=True),
         )
